feat: api logging by pino
